# docs/swagger.yaml
basePath: /
definitions:
...
  common.ApiErr:
    properties:
      Error:
        type: string
    type: object
...
  services.BlockPrice:
    properties:
      blockTime:
        type: integer
      createdAt:
        type: string
      delay:
        type: integer
      id:
        type: integer
      price:
        type: number
    type: object
...
  services.HLDataPriceView:
    properties:
      avgSigns:
        items:
          $ref: '#/definitions/services.HLPriceView'
        type: array
      bigPrice:
        type: string
      code:
        description: 合约代码
        type: string
      dataType:
        description: 最高最低价１最高　２最低价
        type: integer
      isMarketOpening:
        description: 股票杠杆币需要这个字段
        type: boolean
      marketOpenTime:
        description: 股票杠杆币需要这个字段
        type: integer
      priceUsd:
        type: number
      sign:
        description: Sign_Hash值由 Timestamp，DataType,Code,BigPrice
        items:
          type: integer
        type: array
      signs:
        items:
          $ref: '#/definitions/services.HLPriceView'
        type: array
      timestamp:
        description: Node string
        type: integer
    type: object
  services.HLPriceView:
    properties:
      bigPrice:
        type: string
      code:
        description: 合约代码
        type: string
      dataType:
        description: 最高最低价１最高　２最低价
        type: integer
      msg:
        description: debug msg
        type: string
      node:
        type: string
      nodeAddress:
        type: string
      priceUsd:
        type: number
      sign:
        description: Sign_Hash值由 Timestamp，DataType,Code,BigPrice
        items:
          type: integer
        type: array
      timestamp:
        type: integer
    type: object
...
info:
  contact:
    email: wuxiaofei@rechaintech.com
    name: 伍晓飞
  description: oracle-api接口文档.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://test.com/index.html
  title: oracle-api
  version: "1.0"
paths:
...
  /pub/dex/ftx_chart_prices/{coin_type}/{count}/{interval}/{timestamp}:
    get:
      consumes:
      - application/json
      description: 获取杠杆btc代币不同时间区间的价格图表信息
      operationId: FtxChartPricesHandler
      parameters:
      - default: btc3x
        description: ftx类型
        enum:
        - mvi2x
        - btc3x
        - eth3x
        - vix3x
        - gold3x
        - gold3s
        - gold3x
        - eur20x
        - ndx10x
        - govt20x
        - mvi2s
        - btc3s
        - eth3s
        - vix3s
        - gold10s
        - eur20s
        - ndx10s
        - govt20s
        in: path
        name: coin_type
        required: true
        type: string
      - default: 10
        description: 获取多少个数据点
        in: path
        name: count
        required: true
        type: integer
      - default: 1
        description: 数据间隔值,表示多少个15分钟, 如:1表示15分钟间隔 2表示30分钟间隔 3表示45分钟间隔 ,96表示1天间隔 ；
        enum:
        - 1
        - 2
        - 3
        - 4
        - 96
        in: path
        name: interval
        required: true
        type: integer
      - default: 1620383144
        description: 当前时间的unix秒数,该字段未使用，仅在云存储上用于标识
        in: path
        name: timestamp
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: stock info
          headers:
            sign:
              description: 签名信息
              type: string
          schema:
            items:
              $ref: '#/definitions/services.BlockPrice'
            type: array
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: 获取杠杆btc代币不同时间区间的价格图表信息
      tags:
      - default
  /pub/dex/ftx_price/{coin_type}/{data_type}/{timestamp}:
    get:
      consumes:
      - application/json
      description: 获取ftx token价格信息
      operationId: FtxPriceSignHandler
      parameters:
      - default: btc3x
        description: ftx类型
        enum:
        - mvi2x
        - usd
        - btc3x
        - eth3x
        - vix3x
        - govt20x
        - gold3x
        - gold3s
        - eur20x
        - ndx10x
        - mvi2s
        - btc3s
        - eth3s
        - vix3s
        - gold10s
        - eur20s
        - ndx10s
        - govt20s
        in: path
        name: coin_type
        required: true
        type: string
      - default: 1
        description: 最高最低价１最高　２最低价 3平均价 4最后价
        enum:
        - 1
        - 2
        - 3
        - 4
        in: path
        name: data_type
        required: true
        type: integer
      - default: 1620383144
        description: 当前时间的unix秒数,该字段未使用，仅在云存储上用于标识
        in: path
        name: timestamp
        type: integer
      - default: 0
        description: 调试
        in: query
        name: debug
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: token price info
          headers:
            sign:
              description: 签名信息
              type: string
          schema:
            $ref: '#/definitions/services.HLDataPriceView'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: 获取ftx token价格信息
      tags:
      - default
...
  /pub/dex/stock_chart_prices/{coin_type}/{count}/{interval}/{timestamp}:
    get:
      consumes:
      - application/json
      description: 获取股票不同时间区间的价格图表信息
      operationId: StockChartPricesHandler
      parameters:
      - default: AAPL
        description: 股票类型
        enum:
        - AAPL
        - TSLA
        in: path
        name: coin_type
        required: true
        type: string
      - default: 100
        description: 获取多少个数据点
        in: path
        name: count
        required: true
        type: integer
      - default: 4
        description: 数据间隔值,表示多少个15分钟, 如:1表示15分钟间隔 ４表示60分钟间隔 ,96表示1天间隔 ；
        enum:
        - 1
        - 4
        - 96
        in: path
        name: interval
        required: true
        type: integer
      - default: 1620383144
        description: 当前时间的unix秒数,该字段未使用，仅在云存储上用于标识
        in: path
        name: timestamp
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: stock info
          headers:
            sign:
              description: 签名信息
              type: string
          schema:
            items:
              $ref: '#/definitions/services.BlockPrice'
            type: array
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: 获取股票不同时间区间的价格图表信息
      tags:
      - default
...
  /pub/dex/token_price/{token}/{data_type}/{timestamp}:
    get:
      consumes:
      - application/json
      description: 获取token价格信息
      operationId: TokenPriceSignHandler
      parameters:
      - default: 0x66a0f676479cee1d7373f3dc2e2952778bff5bd6
        description: token地址
        in: path
        name: token
        required: true
        type: string
      - default: 1
        description: 最高最低价１最高　２最低价
        enum:
        - 1
        - 2
        in: path
        name: data_type
        required: true
        type: integer
      - default: 1620383144
        description: 当前时间的unix秒数,该字段未使用，仅在云存储上用于标识
        in: path
        name: timestamp
        type: integer
      - default: 0
        description: 调试
        in: query
        name: debug
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: token price info
          headers:
            sign:
              description: 签名信息
              type: string
          schema:
            $ref: '#/definitions/services.HLDataPriceView'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: 获取token价格信息
      tags:
      - default
...
  /pub/internal/dex/ftx_price/{coin_type}/{timestamp}:
    get:
      consumes:
      - application/json
      description: 获取ftx coin最近一小时最高最低价格信息
      operationId: FtxPriceHandler
      parameters:
      - default: btc3x
        description: ftx类型
        enum:
        - mvi2x
        - usd
        - btc3x
        - eth3x
        - vix3x
        - gold3x
        - gold3s
        - eur20x
        - ndx10x
        - govt20x
        - mvi2s
        - btc3s
        - eth3s
        - vix3s
        - gold10s
        - eur20s
        - ndx10s
        - govt20s
        in: path
        name: coin_type
        required: true
        type: string
      - default: 1
        description: 最高最低价１最高　２最低价 3平均价 4最后价
        enum:
        - 1
        - 2
        - 3
        - 4
        in: query
        name: data_type
        required: true
        type: integer
      - default: 1620383144
        description: 当前时间的unix秒数,该字段未使用，仅在云存储上用于标识
        in: path
        name: timestamp
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Price View
          headers:
            sign:
              description: 签名信息
              type: string
          schema:
            $ref: '#/definitions/services.HLPriceView'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: 获取ftx coin最近一小时最高最低价格信息,内部单节点
      tags:
      - default
...
  /pub/internal/token_avgprice:
    post:
      consumes:
      - application/json
      description: 获取token平均价格共识
      operationId: TokenAvgHlPriceHandler
      parameters:
      - description: 节点价格列表
        in: body
        name: nodePrices
        required: true
        schema:
          items:
            $ref: '#/definitions/services.HLPriceView'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: stock info
          headers:
            sign:
              description: 签名信息
              type: string
          schema:
            $ref: '#/definitions/services.HLPriceView'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: '获取token平均价格共识:'
      tags:
      - default
...
swagger: "2.0"